+++ CHANGELOG.md
@@ -1,5 +1,3 @@
 ## Changelog
 
-- Fix a crash on Turkish locales.
-
-[Full Changelog](https://github.com/JamCoreModding/jam-lib/compare/0.6.0+1.20...0.6.1+1.20.x)
+- Fix the mouse scroll event for 1.20.2